feat: implement native app onboarding flow with unit-aware metrics
Add complete 4-step onboarding flow to React Native app matching web
implementation:

  Native App (New):
  - Create main onboarding screen with multi-step navigation and
validation
  - Add progress indicator component showing current step
  - Implement Step 1: Fitness goals (6 options) & experience level (3
options)
  - Implement Step 2: Workout type preferences (6 multi-select options)
  - Implement Step 3: App preferences (weight/distance units, muscle
groups, theme)
  - Implement Step 4: Body metrics (gender, birth year, height, weight -
all optional)
  - Add route protection in drawer layout to redirect incomplete users
  - Copy validation schemas from web app with all option constants

  UX Improvements (Both Apps):
  - Reorder steps: preferences before metrics (users select units first)
  - Height/weight inputs now display in user's preferred units (lbs/kg,
in/cm)
  - Automatic conversion: users enter values in preferred units, stored
as metric
  - Use local state to prevent conversion-on-keystroke issues
  - Dynamic labels show selected units (e.g., "Height (in)" or "Height
(cm)")

  Database & API:
  - Change heightCm and weightKg columns from integer to real for
decimal precision
  - Update validation schemas to accept decimal values
  - Generate migration 0003_magenta_galactus.sql for schema changes

  Technical Details:
  - Native: Uses useState for form state, heroui-native components, Zod
validation
  - Conversions: cm ↔ inches, kg ↔ lbs (rounded to 1 decimal place)
  - Blur-based conversion: convert only when user finishes typing
(prevents React hooks
  violations)
  - Both apps maintain matching behavior and validation logic

+++ packages/api/src/routers/preferences.ts
@@ -22,8 +22,8 @@ const preferencesInput = z.object({
   ).optional(),
   gender: z.enum(["male", "female", "other", "prefer_not_to_say"]).optional(),
   birthYear: z.number().int().min(1920).max(2020).optional(),
-  heightCm: z.number().int().min(50).max(300).optional(),
-  weightKg: z.number().int().min(20).max(500).optional(),
+  heightCm: z.number().min(50).max(300).optional(),
+  weightKg: z.number().min(20).max(500).optional(),
   onboardingCompleted: z.boolean().optional(),
 });
